chatrooms now have authorId when created and can only be deleted by their creators

+++ src/routes/chatrooms.js
@@ -1,5 +1,4 @@
 import express from "express"
-import crypto from "crypto"
 import { db } from "../database.js"
 import { createChatroom, getChatroomById, deleteChatroom, verifyChatroomPassword } from "../database/chatrooms.js"
 import { getAllMessagesByChatroom, getAllMessagesWithUsernames } from "../database/messages.js"
@@ -9,10 +8,9 @@ export const router = express.Router()
 router.post("/create-chatroom", async (req, res) => {
     const name = req.body.name
     const password = req.body.password
-    //const authorId = req.body.user.id
+    const authorId = res.locals.user.id
 
-    
-    await createChatroom(name, password)
+    await createChatroom(name, password, authorId)
     res.redirect("/")
 })
 
@@ -41,17 +39,27 @@ router.post("/delete-chatroom/:id", async (req, res) => {
 */
 router.get("/delete-chatroom/:id", async (req, res) => {
     const idToDelete = Number(req.params.id)
-
     const chatroom = await db("chatrooms").where("id", idToDelete).first()
-
+    
     if(!chatroom) {
-      console.log("Chatroom not found");
-      return res.redirect("back");
+        console.log("Chatroom not found");
+        return res.redirect("back");
+    }
+    const userId = res.locals.user.id
+    const chatroomAuthorId = chatroom.authorId
+
+    if (chatroomAuthorId === userId) {
+        await deleteChatroom(idToDelete)
+        
+    } else {
+        return res.status(400).render("400", {
+            error: "Chatroom can only be deleted by its creator"
+        })
+    
     }
     
-    await deleteChatroom(idToDelete)
-    
     res.redirect("/")
+    
 })
